% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{mb_change_row}
\alias{mb_change_row}
\title{One row of the change table}
\usage{
mb_change_row(
  condition = character(0),
  vocab_id = character(0),
  change = character(0),
  item = character(0),
  old = character(0),
  new = character(0)
)
}
\description{
One row of the change table
}
\keyword{internal}
