% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\docType{data}
\name{MB_CODELIST_COLS}
\alias{MB_CODELIST_COLS}
\title{Columns a code list is written out with, in order.}
\format{
An object of class \code{character} of length 10.
}
\usage{
MB_CODELIST_COLS
}
\description{
Columns a code list is written out with, in order.
}
\keyword{internal}
