% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{mb_exclusion_rules}
\alias{mb_exclusion_rules}
\title{Prior's three rules, mapped to bundled condition keys}
\usage{
mb_exclusion_rules()
}
\description{
Prior's three rules, mapped to bundled condition keys
}
\keyword{internal}
