% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mb_finish_run}
\alias{mb_finish_run}
\title{Shape the return value the same way for every extract_* function}
\usage{
mb_finish_run(run, outdir, verbose)
}
\description{
Shape the return value the same way for every extract_* function
}
\keyword{internal}
