% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{mb_diagnose_one}
\alias{mb_diagnose_one}
\title{Extract one diagnosis-based condition}
\usage{
mb_diagnose_one(
  lpr,
  condition,
  codes,
  id_col,
  code_col,
  date_col,
  from,
  to,
  dedupe_same_day = TRUE,
  keep_events = FALSE,
  ids = NULL
)
}
\description{
Extract one diagnosis-based condition
}
\keyword{internal}
